Move new instructions to its own module
